Not working because:
-although Fishbench shows no slowdown there is a slowdown (worth -7 ELO)
-context is already reflected by the bonus decreasings

+++ src/types.h
@@ -194,7 +194,7 @@ enum Value : int {
 
 struct ValPly {
   Value val;
-  float gameply;
+  int gameply;
 };
 
 enum PieceType {
